+++ src/main/java/it/gov/pagopa/payhub/ionotification/config/rest/IORestConnectorConfig.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.ionotification.config;
+package it.gov.pagopa.payhub.ionotification.config.rest;
 
 import it.gov.pagopa.payhub.ionotification.connector.io.IOFeignRestClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
