+++ src/test/java/it/gov/pagopa/payhub/ionotification/utils/HttpUtilsTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.payhub.ionotification.utils;
 
-import it.gov.pagopa.payhub.ionotification.config.HttpClientConfig;
+import it.gov.pagopa.payhub.ionotification.config.rest.HttpClientConfig;
 import org.apache.hc.client5.http.HttpRoute;
 import org.apache.hc.client5.http.classic.HttpClient;
 import org.apache.hc.client5.http.config.ConnectionConfig;
